Break out deposit Pure module in Pure.State....
